refactor(arenabench): split the transcript reader out of telemetry.py
telemetry.py crossed the 1500-line ratchet in 76c7aae (#2391) at 1534 lines,
turning the `file size ratchet` check red on main and, through the merge
result, on every open PR.

The seam is not arbitrary. The transcript half shares *no* module-level name
with the metrics half -- verified by name, not by eye: no path constant, no
reader, no reducer, nothing. The two halves answer different questions about
the same artifacts (reduce a trial to the scoreboard's dimensions, vs. replay
one trial's stream for a human reading along), and were one file by habit.

  telemetry.py   1534 -> 1121
  transcript.py          437  (new)

No re-export shim: `TranscriptReader` moves and its four importers move with
it (server.py, monitor.py's doc reference, test_proof.py, test_telemetry.py).
A compatibility alias inside a package that is being prepared for ejection
would be indirection with no consumer.

The guard's own gap -- the base-relative allowance from #2004 never runs for a
first-time crossing, which is why this failed PRs that did not touch the file
-- is #2397 and is deliberately not fixed here.

arenabench pytest: 579 passed, 6 skipped.

Refs #2397
